Mostly cosmetic changes for ccf related scripts
Fixed spelling issue, removed commented code, and
(only functional change) added one more check for
port availability before starting the CCF services.

+++ ledgers/ccf/scripts/start_ccf_network.sh
@@ -58,6 +58,12 @@ if [ -f ${PDO_HOME}/ccf/workspace/sandbox_0/node.pid ]; then
     fi
 fi
 
+if lsof -Pi "@${F_INTERFACE}:${F_PORT}" -sTCP:LISTEN -t > /dev/null
+then
+    yell interface "local:${F_INTERFACE}:${F_PORT}" appears to be in use
+    exit -1
+fi
+
 rm -f ${PDO_LEDGER_KEY_ROOT}/networkcert.pem
 rm -f ${PDO_LEDGER_KEY_ROOT}/ledger_authority_pub.pem
